+++ Paseto.Net.sln
@@ -1,13 +1,14 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paseto", "Paseto\Paseto.csproj", "{980FA4A3-4DE9-4379-915D-583FCD65B129}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Paseto", "Paseto\Paseto.csproj", "{980FA4A3-4DE9-4379-915D-583FCD65B129}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paseto.Tests", "Paseto.Tests\Paseto.Tests.csproj", "{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Paseto.Tests", "Paseto.Tests\Paseto.Tests.csproj", "{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paseto.Public", "Paseto.Public\Paseto.Public.csproj", "{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Paseto.Public", "Paseto.Public\Paseto.Public.csproj", "{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Paseto.Public.Tests", "Paseto.Public.Tests\Paseto.Public.Tests.csproj", "{E0EB6703-F531-4C23-B949-101C3AEF8FCE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -18,45 +19,60 @@ Global
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|x64.ActiveCfg = Debug|x64
-		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|x64.Build.0 = Debug|x64
-		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|x86.ActiveCfg = Debug|x86
-		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|x86.Build.0 = Debug|x86
+		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|x64.Build.0 = Debug|Any CPU
+		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Debug|x86.Build.0 = Debug|Any CPU
 		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|Any CPU.Build.0 = Release|Any CPU
-		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|x64.ActiveCfg = Release|x64
-		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|x64.Build.0 = Release|x64
-		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|x86.ActiveCfg = Release|x86
-		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|x86.Build.0 = Release|x86
+		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|x64.ActiveCfg = Release|Any CPU
+		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|x64.Build.0 = Release|Any CPU
+		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|x86.ActiveCfg = Release|Any CPU
+		{980FA4A3-4DE9-4379-915D-583FCD65B129}.Release|x86.Build.0 = Release|Any CPU
 		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|x64.ActiveCfg = Debug|x64
-		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|x64.Build.0 = Debug|x64
-		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|x86.ActiveCfg = Debug|x86
-		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|x86.Build.0 = Debug|x86
+		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|x64.Build.0 = Debug|Any CPU
+		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Debug|x86.Build.0 = Debug|Any CPU
 		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|x64.ActiveCfg = Release|x64
-		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|x64.Build.0 = Release|x64
-		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|x86.ActiveCfg = Release|x86
-		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|x86.Build.0 = Release|x86
+		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|x64.ActiveCfg = Release|Any CPU
+		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|x64.Build.0 = Release|Any CPU
+		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|x86.ActiveCfg = Release|Any CPU
+		{AD714AF1-4EB9-45B0-9C8B-25615B4F70B9}.Release|x86.Build.0 = Release|Any CPU
 		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|x64.ActiveCfg = Debug|x64
-		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|x64.Build.0 = Debug|x64
-		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|x86.ActiveCfg = Debug|x86
-		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|x86.Build.0 = Debug|x86
+		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|x64.Build.0 = Debug|Any CPU
+		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Debug|x86.Build.0 = Debug|Any CPU
 		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|x64.ActiveCfg = Release|x64
-		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|x64.Build.0 = Release|x64
-		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|x86.ActiveCfg = Release|x86
-		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|x86.Build.0 = Release|x86
+		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|x64.ActiveCfg = Release|Any CPU
+		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|x64.Build.0 = Release|Any CPU
+		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|x86.ActiveCfg = Release|Any CPU
+		{2FD798CB-3ACA-4BAD-8AA6-D39EDADEC628}.Release|x86.Build.0 = Release|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Debug|x64.Build.0 = Debug|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Debug|x86.Build.0 = Debug|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Release|x64.ActiveCfg = Release|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Release|x64.Build.0 = Release|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Release|x86.ActiveCfg = Release|Any CPU
+		{E0EB6703-F531-4C23-B949-101C3AEF8FCE}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {92832FE5-A325-4A7E-9D0E-79C464D2492E}
 	EndGlobalSection
 EndGlobal
